# [PoC] Change Processor.OnEmit to return record passed to next registered processor