--- a/sdk/log/batch_test.go
+++ b/sdk/log/batch_test.go
@@ -46,7 +46,10 @@ func TestEmptyBatchConfig(t *testing.T) {
 		var bp BatchProcessor
 		ctx := context.Background()
 		var record Record
-		assert.NoError(t, bp.OnEmit(ctx, record), "OnEmit")
+		record.SetBody(log.StringValue("message"))
+		r, err := bp.OnEmit(ctx, record)
+		assert.Equal(t, record, r, "OnEmit record")
+		assert.NoError(t, err, "OnEmit err")
 		assert.False(t, bp.Enabled(ctx, record), "Enabled")
 		assert.NoError(t, bp.ForceFlush(ctx), "ForceFlush")
 		assert.NoError(t, bp.Shutdown(ctx), "Shutdown")
@@ -198,7 +201,8 @@ func TestBatchProcessor(t *testing.T) {
 			WithExportTimeout(time.Hour),
 		)
 		for _, r := range make([]Record, size) {
-			assert.NoError(t, b.OnEmit(ctx, r))
+			_, err := b.OnEmit(ctx, r)
+			assert.NoError(t, err)
 		}
 		var got []Record
 		assert.Eventually(t, func() bool {
@@ -221,7 +225,8 @@ func TestBatchProcessor(t *testing.T) {
 			WithExportTimeout(time.Hour),
 		)
 		for _, r := range make([]Record, 10*batch) {
-			assert.NoError(t, b.OnEmit(ctx, r))
+			_, err := b.OnEmit(ctx, r)
+			assert.NoError(t, err)
 		}
 		assert.Eventually(t, func() bool {
 			return e.ExportN() > 1
@@ -244,7 +249,8 @@ func TestBatchProcessor(t *testing.T) {
 			WithExportTimeout(time.Hour),
 		)
 		for _, r := range make([]Record, 2*batch) {
-			assert.NoError(t, b.OnEmit(ctx, r))
+			_, err := b.OnEmit(ctx, r)
+			assert.NoError(t, err)
 		}
 
 		var n int
@@ -255,7 +261,7 @@ func TestBatchProcessor(t *testing.T) {
 
 		var err error
 		require.Eventually(t, func() bool {
-			err = b.OnEmit(ctx, Record{})
+			_, err = b.OnEmit(ctx, Record{})
 			return true
 		}, time.Second, time.Microsecond, "OnEmit blocked")
 		assert.NoError(t, err)
@@ -303,7 +309,8 @@ func TestBatchProcessor(t *testing.T) {
 			assert.NoError(t, b.Shutdown(ctx))
 
 			want := e.ExportN()
-			assert.NoError(t, b.OnEmit(ctx, Record{}))
+			_, err := b.OnEmit(ctx, Record{})
+			assert.NoError(t, err)
 			assert.Equal(t, want, e.ExportN(), "Export called after shutdown")
 		})
 
@@ -311,7 +318,8 @@ func TestBatchProcessor(t *testing.T) {
 			e := newTestExporter(nil)
 			b := NewBatchProcessor(e)
 
-			assert.NoError(t, b.OnEmit(ctx, Record{}))
+			_, err := b.OnEmit(ctx, Record{})
+			assert.NoError(t, err)
 			assert.NoError(t, b.Shutdown(ctx))
 
 			assert.NoError(t, b.ForceFlush(ctx))
@@ -346,7 +354,8 @@ func TestBatchProcessor(t *testing.T) {
 
 			var r Record
 			r.SetBody(log.BoolValue(true))
-			require.NoError(t, b.OnEmit(ctx, r))
+			_, err := b.OnEmit(ctx, r)
+			require.NoError(t, err)
 
 			assert.ErrorIs(t, b.ForceFlush(ctx), assert.AnError, "exporter error not returned")
 			assert.Equal(t, 1, e.ForceFlushN(), "exporter ForceFlush calls")
@@ -381,7 +390,8 @@ func TestBatchProcessor(t *testing.T) {
 
 			// Enqueue 10 x "batch size" amount of records.
 			for i := 0; i < 10*batch; i++ {
-				require.NoError(t, b.OnEmit(ctx, Record{}))
+				_, err := b.OnEmit(ctx, Record{})
+				require.NoError(t, err)
 			}
 			assert.Eventually(t, func() bool {
 				return e.ExportN() > 0 && len(b.exporter.input) == cap(b.exporter.input)
@@ -425,7 +435,7 @@ func TestBatchProcessor(t *testing.T) {
 
 			var r Record
 			r.SetBody(log.BoolValue(true))
-			_ = b.OnEmit(ctx, r)
+			_, _ = b.OnEmit(ctx, r)
 			t.Cleanup(func() { _ = b.Shutdown(ctx) })
 			t.Cleanup(func() { close(e.ExportTrigger) })
 
@@ -455,21 +465,26 @@ func TestBatchProcessor(t *testing.T) {
 		)
 		var r Record
 		// First record will be blocked by testExporter.Export
-		assert.NoError(t, b.OnEmit(ctx, r), "exported record")
+
+		_, err := b.OnEmit(ctx, r)
+		assert.NoError(t, err, "exported record")
 		require.Eventually(t, func() bool {
 			return e.ExportN() > 0
 		}, 2*time.Second, time.Microsecond, "blocked export not attempted")
 
 		// Second record will be written to export queue
-		assert.NoError(t, b.OnEmit(ctx, r), "export queue record")
+		_, err = b.OnEmit(ctx, r)
+		assert.NoError(t, err, "export queue record")
 		require.Eventually(t, func() bool {
 			return len(b.exporter.input) == cap(b.exporter.input)
 		}, 2*time.Second, time.Microsecond, "blocked queue read not attempted")
 
 		// Third record will be written to BatchProcessor.q
-		assert.NoError(t, b.OnEmit(ctx, r), "first queued")
+		_, err = b.OnEmit(ctx, r)
+		assert.NoError(t, err, "first queued")
 		// The previous record will be dropped, as the new one will be written to BatchProcessor.q
-		assert.NoError(t, b.OnEmit(ctx, r), "second queued")
+		_, err = b.OnEmit(ctx, r)
+		assert.NoError(t, err, "second queued")
 
 		wantMsg := `"level"=1 "msg"="dropped log records" "dropped"=1`
 		assert.Eventually(t, func() bool {
@@ -497,7 +512,8 @@ func TestBatchProcessor(t *testing.T) {
 					case <-ctx.Done():
 						return
 					default:
-						assert.NoError(t, b.OnEmit(ctx, Record{}))
+						_, err := b.OnEmit(ctx, Record{})
+						assert.NoError(t, err)
 						// Ignore partial flush errors.
 						_ = b.ForceFlush(ctx)
 					}
@@ -663,7 +679,7 @@ func BenchmarkBatchProcessorOnEmit(b *testing.B) {
 	b.RunParallel(func(pb *testing.PB) {
 		var err error
 		for pb.Next() {
-			err = bp.OnEmit(ctx, r)
+			r, err = bp.OnEmit(ctx, r)
 		}
 		_ = err
 	})